Trigger internal build

+++ pysasktran/sasktran/test_emissions.py
@@ -29,9 +29,9 @@ class TestEmission(unittest.TestCase):
         geometry.from_sza_saa(sza=60, saa=60, lat=0, lon=0, tanalts_km=[10, 20, 30, 40], mjd=54372, locallook=0,
                               satalt_km=600, refalt_km=20)
 
-        alts = np.arange(500, 100500, 1000)
-        wavel = np.linspace(300, 350, 50)
-        ver = np.ones((len(alts), len(wavel)))
+        alts = np.arange(500, 100500, 1000, dtype='float')
+        wavel = np.linspace(300, 350, 50, dtype='float')
+        ver = np.ones((len(alts), len(wavel)), dtype='float')
         emission = sk.EmissionTable(alts, wavel, ver)
 
         atmosphere = sk.Atmosphere()
